refactor: update businesslist components code
